python/requirements.txt
pymupdf
pdfminer.six
sumy
requests
nltk
